+++ src/ChoyUtilities/Assets/ChoyUtilities/Common/Runtime/Core/Exception/FloaterException.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 
 namespace ChoyUtilities {
-    public class FloaterException : Exception {
+    public sealed class FloaterException : Exception {
         public FloaterException(string message) : base(message) { }
     }
 }