+++ src/ChoyUtilities/Assets/ChoyUtilities/Common/Runtime/Core/Exception/RawSetException.cs
@@ -1,0 +1,8 @@
+﻿using System;
+
+namespace ChoyUtilities {
+    
+    public sealed class RawSetException : Exception {
+        public RawSetException(string message) : base(message) { }
+    }
+}