+++ src/ChoyUtilities/Assets/ChoyUtilities/Common/Runtime/Core/Exception/SingletonException.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 
 namespace ChoyUtilities {
-    public class SingletonException : Exception {
+    public sealed class SingletonException : Exception {
         public SingletonException(string message) : base(message) { }
     }
 }